white_box/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
cryptography==46.0.1
pymysql==1.1.1
pytest==9.0.2
python-dotenv==1.2.2
